backend/requirements.txt
flask
flask-cors
python-dotenv
flask-sqlalchemy
PyJWT
werkzeug
